pyproject.toml
[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"

[project]
name = "DOLFINx-ADJOINT"
version = "0.4.0"
authors = [
  { name = "Niklas Hornischer", email = "niklas.hornischer@ians.uni-stuttgart.de" },
]
maintainers = [
  { name = "Niklas Hornischer", email = "niklas.hornischer@ians.uni-stuttgart.de" },
]

description = "Automatic differentiation library for DOLFINx"
readme = "README.md"
requires-python = ">=3.12"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
license = { file = "LICENSE.txt" }
dependencies = [
  "mpi4py",
  "numpy",
  "scipy",
  "fenics-dolfinx>=0.10.0",
]

[project.optional-dependencies]
test = ["pytest", "networkx", "gmsh", "isort", "black"]
demos = ["networkx", "gmsh"]
all = ["pytest", "networkx", "matplotlib", "gmsh", "isort", "black"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]